+++ src/rics/__init__.py
@@ -18,6 +18,6 @@ __all__ = [
     "configure_stuff",
 ]
 
-__version__ = "6.0.1.dev1"
+__version__ = "6.0.2"
 
 _logging.getLogger(__name__).addHandler(_logging.NullHandler())